Fix runtime bugs, add error messages, fix desgin
Some error with the write and operate functions were fixed, name of
variables were changed to make the code more readable. Functions into
the code were ordered to make it easier to understand. Comments were
added to make it is easier to know  the function of each section.
The design is more compact and colors are more brilliant.